cds/scripts/requirements.txt
pandas==2.0.0
psycopg2-binary==2.9.5
sqlalchemy==1.4.46
requests==2.28.2
python-dotenv==0.21.1 